refactor: extract UI components and options
