+++ src/App.jsx
@@ -1,257 +1,27 @@
 import React, { useMemo, useState } from "react";
-import { Copy, Shuffle, RotateCcw, Download, Languages, Wand2, ShieldCheck, CameraOff, EyeOff, Sliders } from "lucide-react";
-
-// ===== UI PRIMITIVES (no external UI deps) =====
-const Card = ({ children, className = "" }) => (
-  <div className={`rounded-2xl shadow-sm border border-gray-200 bg-white ${className}`}>{children}</div>
-);
-const CardHeader = ({ title, subtitle, right }) => (
-  <div className="px-5 pt-5 pb-3 border-b flex items-start justify-between gap-3">
-    <div>
-      <h3 className="text-lg font-semibold tracking-tight">{title}</h3>
-      {subtitle && <p className="text-sm text-gray-500 mt-1">{subtitle}</p>}
-    </div>
-    {right}
-  </div>
-);
-const CardContent = ({ children, className = "" }) => (
-  <div className={`px-5 py-4 ${className}`}>{children}</div>
-);
-const Button = ({ children, onClick, className = "", variant = "default", title, type = "button" }) => {
-  const base = "inline-flex items-center gap-2 rounded-2xl px-3.5 py-2.5 text-sm font-medium transition border focus:outline-none";
-  const styles = {
-    default: "bg-black text-white border-black hover:opacity-90",
-    ghost: "bg-white text-gray-700 border-gray-200 hover:bg-gray-50",
-    subtle: "bg-gray-900/5 text-gray-800 border-gray-200 hover:bg-gray-900/10",
-  };
-  return (
-    <button type={type} title={title} onClick={onClick} className={`${base} ${styles[variant]} ${className}`}>
-      {children}
-    </button>
-  );
-};
-const Select = ({ value, onChange, options, className = "" }) => (
-  <select value={value} onChange={(e) => onChange(e.target.value)} className={`w-full rounded-xl border border-gray-300 bg-white px-3 py-2 text-sm ${className}`}>
-    {options.map((o) => (
-      <option key={o.value} value={o.value}>
-        {o.label}
-      </option>
-    ))}
-  </select>
-);
-const Input = ({ value, onChange, placeholder, className = "", type = "text", min, max, step }) => (
-  <input
-    value={value}
-    onChange={(e) => onChange(type === "number" ? Number(e.target.value) : e.target.value)}
-    placeholder={placeholder}
-    type={type}
-    min={min}
-    max={max}
-    step={step}
-    className={`w-full rounded-xl border border-gray-300 bg-white px-3 py-2 text-sm ${className}`}
-  />
-);
-const Textarea = ({ value, onChange, placeholder, rows = 4, className = "" }) => (
-  <textarea value={value} onChange={(e) => onChange(e.target.value)} placeholder={placeholder} rows={rows} className={`w-full rounded-xl border border-gray-300 bg-white px-3 py-2 text-sm ${className}`} />
-);
-const Toggle = ({ checked, onChange, label }) => (
-  <label className="flex items-center gap-2 text-sm">
-    <input type="checkbox" checked={checked} onChange={(e) => onChange(e.target.checked)} className="h-4 w-4" />
-    <span>{label}</span>
-  </label>
-);
-
-// ===== OPTIONS (EN + JP) =====
-const options = {
-  age: [
-    { en: "teen", jp: "10代" },
-    { en: "young adult", jp: "若い大人" },
-    { en: "adult", jp: "大人" },
-    { en: "mature", jp: "成熟した大人" },
-    { en: "child", jp: "子供" },
-    { en: "senior", jp: "高齢者" },
-  ],
-  gender: [
-    { en: "female", jp: "女性" },
-    { en: "male", jp: "男性" },
-    { en: "non-binary", jp: "ノンバイナリー" },
-    { en: "agender", jp: "無性別" },
-  ],
-  ethnicity: [
-    { en: "Japanese", jp: "日本人風" },
-    { en: "East Asian", jp: "東アジア風" },
-    { en: "Mixed", jp: "ミックス" },
-    { en: "Caucasian", jp: "白人風" },
-    { en: "Latina", jp: "ラティーナ風" },
-    { en: "Black", jp: "黒人風" },
-    { en: "South Asian", jp: "南アジア風" },
-    { en: "Southeast Asian", jp: "東南アジア風" },
-    { en: "Middle Eastern", jp: "中東風" },
-  ],
-  face: [
-    { en: "natural features", jp: "自然な顔立ち" },
-    { en: "soft oval face", jp: "柔らかな卵型の輪郭" },
-    { en: "defined jawline", jp: "輪郭のはっきりした顎" },
-    { en: "light freckles", jp: "薄いそばかす" },
-    { en: "beauty mark under eye", jp: "目の下のほくろ" },
-    { en: "subtle dimples", jp: "控えめなえくぼ" },
-    { en: "sharp cheekbones", jp: "シャープな頬骨" },
-    { en: "almond eyes", jp: "アーモンド形の目" },
-    { en: "soft blush on cheeks", jp: "頬の淡い赤み" },
-    { en: "upturned nose", jp: "上向きの鼻" },
-    { en: "sleepy hooded eyes", jp: "眠そうな一重まぶた" },
-  ],
-  hairStyle: [
-    { en: "long straight", jp: "ロングのストレート" },
-    { en: "long wavy", jp: "ロングの緩いウェーブ" },
-    { en: "low ponytail", jp: "低い位置のポニーテール" },
-    { en: "bob", jp: "ボブ" },
-    { en: "side bangs", jp: "サイドバング" },
-    { en: "pixie cut", jp: "ピクシーカット" },
-    { en: "top knot", jp: "トップノット" },
-    { en: "messy bun", jp: "ラフなお団子" },
-    { en: "curly bob", jp: "カールのボブ" },
-    { en: "double braids", jp: "二つ編み" },
-  ],
-  makeup: [
-    { en: "none", jp: "メイクなし" },
-    { en: "natural makeup", jp: "ナチュラルメイク" },
-    { en: "soft glam", jp: "ソフトグラム" },
-    { en: "bold makeup", jp: "大胆なメイク" },
-    { en: "smokey eye", jp: "スモーキーアイ" },
-    { en: "glossy lips", jp: "ツヤのあるリップ" },
-    { en: "cat-eye eyeliner", jp: "キャットアイライン" },
-  ],
-  eyeColor: [
-    { en: "dark brown", jp: "ダークブラウン" },
-    { en: "brown", jp: "ブラウン" },
-    { en: "hazel", jp: "ヘーゼル" },
-    { en: "black", jp: "黒" },
-    { en: "gray", jp: "グレー" },
-    { en: "blue", jp: "ブルー" },
-    { en: "green", jp: "グリーン" },
-  ],
-  tops: [
-    { en: "fitted light gray long-sleeve, slightly open neckline", jp: "やや胸元の開いた薄いグレーの長袖" },
-    { en: "oversized knit sweater", jp: "オーバーサイズのニット" },
-    { en: "simple white tee", jp: "シンプルな白T" },
-    { en: "button-up shirt", jp: "ボタンアップシャツ" },
-    { en: "casual hoodie", jp: "カジュアルなパーカー" },
-    { en: "silk blouse", jp: "シルクのブラウス" },
-    { en: "graphic tee", jp: "グラフィックT" },
-    { en: "turtleneck sweater", jp: "タートルネックセーター" },
-  ],
-  bottoms: [
-    { en: "dark trousers", jp: "ダークトラウザー" },
-    { en: "pleated skirt", jp: "プリーツスカート" },
-    { en: "high-waist jeans", jp: "ハイウエストジーンズ" },
-    { en: "tailored shorts", jp: "テーラードショーツ" },
-    { en: "pencil skirt", jp: "タイトスカート" },
-    { en: "flowy maxi skirt", jp: "ゆったりしたマキシスカート" },
-    { en: "ripped jeans", jp: "ダメージジーンズ" },
-  ],
-  outer: [
-    { en: "", jp: "" },
-    { en: "tailored blazer", jp: "テーラードブレザー" },
-    { en: "light cardigan", jp: "薄手のカーディガン" },
-    { en: "denim jacket", jp: "デニムジャケット" },
-    { en: "trench coat", jp: "トレンチコート" },
-    { en: "leather jacket", jp: "レザージャケット" },
-    { en: "puffer jacket", jp: "中綿ジャケット" },
-    { en: "hooded coat", jp: "フード付きコート" },
-  ],
-  accessories: [
-    { en: "round glasses", jp: "丸メガネ" },
-    { en: "thin necklace", jp: "細いネックレス" },
-    { en: "stud earrings", jp: "スタッドピアス" },
-    { en: "headphones", jp: "ヘッドホン" },
-    { en: "watch", jp: "腕時計" },
-    { en: "no accessories", jp: "アクセサリーなし" },
-    { en: "scarf", jp: "スカーフ" },
-    { en: "bracelet", jp: "ブレスレット" },
-    { en: "beanie hat", jp: "ビーニー帽" },
-    { en: "backpack", jp: "バックパック" },
-    { en: "fingerless gloves", jp: "指なし手袋" },
-  ],
-  background: [
-    { en: "minimal studio backdrop", jp: "最小限のスタジオ背景" },
-    { en: "cozy room with large window", jp: "大きな窓のある居心地のよい部屋" },
-    { en: "modern café interior", jp: "モダンなカフェの内装" },
-    { en: "night city through window", jp: "窓越しの夜の街並み" },
-    { en: "library with tall shelves", jp: "高い本棚のある図書館" },
-    { en: "sunlit park", jp: "日差しのある公園" },
-    { en: "rain-soaked street at night", jp: "夜の雨に濡れた街路" },
-    { en: "mountain cabin interior", jp: "山小屋の室内" },
-    { en: "traditional tatami room", jp: "和室の畳" },
-  ],
-  bgDetails: [
-    { en: "rain reflections on glass", jp: "ガラスの雨反射" },
-    { en: "soft clutter of books and notes", jp: "本とノートの柔らかな雑然" },
-    { en: "neon glow spill", jp: "ネオンの光漏れ" },
-    { en: "indoor plants", jp: "観葉植物" },
-    { en: "wooden desk", jp: "木製デスク" },
-    { en: "retro posters", jp: "レトロポスター" },
-    { en: "hanging fairy lights", jp: "吊るされたフェアリーライト" },
-    { en: "steam rising from mug", jp: "マグから立ち上る湯気" },
-    { en: "floating dust in sunlight", jp: "日差しに舞う塵" },
-    { en: "stack of vinyl records", jp: "積み重なったレコード" },
-    { en: "old CRT monitor glow", jp: "古いCRTモニターの光" },
-  ],
-  activity: [
-    { en: "studying and writing in a notebook", jp: "ノートに勉強・筆記" },
-    { en: "reading a paperback", jp: "文庫本を読む" },
-    { en: "using a smartphone thoughtfully", jp: "スマホを思慮深く操作" },
-    { en: "sipping coffee while thinking", jp: "コーヒーを飲み考え込む" },
-    { en: "staring out of the window", jp: "窓の外を見つめる" },
-    { en: "typing on a laptop", jp: "ノートPCでタイピング" },
-    { en: "sketching in a notebook", jp: "ノートにスケッチする" },
-    { en: "playing acoustic guitar", jp: "アコースティックギターを弾く" },
-    { en: "painting on a canvas", jp: "キャンバスに絵を描く" },
-    { en: "stretching arms lazily", jp: "腕を伸ばして伸びをする" },
-  ],
-  shot: [
-    { en: "medium shot", jp: "ミディアムショット" },
-    { en: "medium close-up", jp: "ミディアムクローズアップ" },
-    { en: "close-up", jp: "クローズアップ" },
-    { en: "long shot", jp: "ロングショット" },
-    { en: "wide shot", jp: "ワイドショット" },
-    { en: "full body shot", jp: "全身ショット" },
-    { en: "extreme close-up", jp: "エクストリームクローズアップ" },
-    { en: "over-the-shoulder shot", jp: "肩越しショット" },
-  ],
-  lighting: [
-    { en: "soft desk lamp + ambient practicals", jp: "柔らかなデスクランプ＋室内の実用光" },
-    { en: "cinematic low-key lighting", jp: "シネマ調のローキー照明" },
-    { en: "sunset window light", jp: "夕暮れの窓明かり" },
-    { en: "bright overcast daylight", jp: "明るい曇天の昼光" },
-    { en: "neon sign lighting", jp: "ネオン看板の光" },
-    { en: "warm fireplace glow", jp: "暖炉の暖かい光" },
-    { en: "cool moonlight through window", jp: "窓から差し込む冷たい月光" },
-    { en: "dramatic chiaroscuro", jp: "劇的なキアロスクーロ" },
-  ],
-  mood: [
-    { en: "calm and focused", jp: "穏やかで集中" },
-    { en: "quietly reflective", jp: "静かに物思い" },
-    { en: "warm and intimate", jp: "温かく親密" },
-    { en: "energetic and lively", jp: "エネルギッシュで活気がある" },
-    { en: "melancholic", jp: "物悲しい" },
-    { en: "nostalgic", jp: "懐かしい" },
-    { en: "mysterious", jp: "神秘的" },
-    { en: "joyful", jp: "喜びに満ちた" },
-  ],
-  style: [
-    { en: "photorealistic, ultra-detailed, natural skin texture", jp: "超写実・精緻・自然な肌感" },
-    { en: "cinematic color grading", jp: "シネマ調カラーグレーディング" },
-    { en: "anime-inspired clean lines", jp: "アニメ風のクリーンな線" },
-    { en: "vintage film look", jp: "ビンテージフィルム風" },
-    { en: "soft film grain", jp: "柔らかなフィルムグレイン" },
-    { en: "washed-out 90s photo", jp: "色あせた90年代写真" },
-    { en: "VHS tape artifacts", jp: "VHSテープのノイズ" },
-    { en: "handheld camcorder style", jp: "ハンディカム風" },
-  ],
-};
-
-const toSelectOptions = (arr) => arr.map((i) => ({ value: i.en, label: i.en }));
+import {
+  Copy,
+  Shuffle,
+  RotateCcw,
+  Download,
+  Languages,
+  Wand2,
+  ShieldCheck,
+  CameraOff,
+  EyeOff,
+  Sliders,
+} from "lucide-react";
+import {
+  Card,
+  CardHeader,
+  CardContent,
+  Button,
+  Select,
+  Input,
+  Textarea,
+  Toggle,
+} from "./components/ui";
+import { options, toSelectOptions, findJP } from "./data/options";
 
 function useClipboard() {
   const copy = async (text) => {
@@ -265,13 +35,6 @@ function useClipboard() {
   return { copy };
 }
 
-function findJP(en) {
-  for (const group of Object.values(options)) {
-    const hit = group.find((i) => i.en === en);
-    if (hit) return hit.jp || en;
-  }
-  return en;
-}
 
 // ===== DEFAULT STATE (extended for candid & static camera) =====
 const defaultState = {
